+++ src/devices/WH45.ts
@@ -1,10 +1,11 @@
-import { PlatformAccessory, Service } from 'homebridge';
+import { PlatformAccessory } from 'homebridge';
 import { EcowittPlatform } from './../EcowittPlatform';
 import { EcowittAccessory } from './../EcowittAccessory';
 import { TemperatureSensor } from './../sensors/TemperatureSensor';
 import { HumiditySensor } from './../sensors/HumiditySensor';
 import { AirQualitySensor } from './../sensors/AirQualitySensor';
 import { CarbonDioxideSensor } from './../sensors/CarbonDioxideSensor';
+import { BatterySensor } from './../sensors/BatterySensor';
 import * as utils from './../Utils';
 
 //------------------------------------------------------------------------------
@@ -14,7 +15,7 @@ export class WH45 extends EcowittAccessory {
     'airQualityPM25Avg', 'airQualityPM10', 'airQualityPM10Avg', 'carbonDioxide',
     'carbonDioxideAvg'];
 
-  protected battery: Service;
+  protected battery: BatterySensor | undefined;
   protected temperature: TemperatureSensor | undefined;
   protected humidity: HumiditySensor | undefined;
   protected airQualityPM25: AirQualitySensor | undefined;
@@ -33,12 +34,19 @@ export class WH45 extends EcowittAccessory {
     this.requiredData = ['co2_batt', 'tf_co2', 'humi_co2', 'pm25_co2', 'pm25_24h_co2',
       'pm10_co2', 'pm10_24h_co2', 'co2', 'co2_24h'];
 
-    this.battery = this.addBattery('', true);
-
     const hideConfig = this.platform.config?.hidden || {};
     const hidden = Object.keys(hideConfig).filter(k => !!hideConfig[k]);
 
     let nameOverride: string | undefined;
+
+    if (!utils.includesAny(hidden, ['battery', `${this.shortServiceId}:battery`])) {
+      nameOverride = utils.lookup(this.platform.config?.nameOverrides, `${this.shortServiceId}:battery`);
+      this.battery = new BatterySensor(platform, accessory, `${this.accessoryId}:battery`, nameOverride || 'Battery');
+    } else {
+      this.battery = new BatterySensor(platform, accessory, `${this.accessoryId}:battery`, 'Battery');
+      this.battery.removeService();
+      this.battery = undefined;
+    }
 
     if (!utils.includesAny(hidden, ['temperature', `${this.shortServiceId}:temperature`])) {
       nameOverride = utils.lookup(this.platform.config?.nameOverrides, `${this.shortServiceId}:temperature`);
@@ -132,9 +140,20 @@ export class WH45 extends EcowittAccessory {
     const batteryLevel = batt / 5.0;
     const lowBattery = batt <= 1.1;
 
-    this.updateBatteryLevel(this.battery, utils.boundRange(batteryLevel * 100));
-    this.updateStatusLowBattery(this.battery, lowBattery);
-    this.updateChargingState(this.battery, batt === 6);
+    this.battery?.updateLevel(
+      utils.boundRange(batteryLevel * 100),
+      dataReport.dateutc,
+    );
+
+    this.battery?.updateStatusLow(
+      lowBattery,
+      dataReport.dateutc,
+    );
+
+    this.battery?.updateChargingState(
+      batt === 6,
+      dataReport.dateutc,
+    );
 
     this.temperature?.update(
       parseFloat(dataReport['tf_co2']),