+++ tests/unit/devices/WH51.spec.ts
@@ -20,7 +20,11 @@ configs.forEach(config => {
   describe(`WH51 device should be configured ${config}`, () => {
     before('Initialize device', () => {
       platform = createPlatform(config);
-      accessory = new api.platformAccessory('Accessory', "5746853e-4fee-4e47-97dd-53065ef1de03")
+      accessory = new api.platformAccessory('Accessory', "5746853e-4fee-4e47-97dd-53065ef1de03");
+
+      platform.config.nameOverrides = [];
+      platform.config.hidden = {};
+
       device = new WH51(platform, accessory, 1);
     });
 
@@ -32,7 +36,7 @@ configs.forEach(config => {
     it('Services are created', (done) => {
       expect(device.battery).to.not.be.undefined;
       expect(device.soilMoisture).to.not.be.undefined;
-      expect(device.battery.displayName).to.equal('');
+      expect(device.battery.service.displayName).to.equal('Battery');
       expect(device.soilMoisture.service.displayName).to.equal("Soil Moisture");
       done();
     });
@@ -40,8 +44,8 @@ configs.forEach(config => {
     it('Update is called successfully', (done) => {
       device.update(dataReport);
 
-      expect(device.battery.characteristics[0].value).to.equal(0); // low batt
-      expect(device.battery.characteristics[3].value).to.equal(81); // batt percentage
+      expect(device.battery.service.characteristics[1].value).to.equal(0); // low batt
+      expect(device.battery.service.characteristics[4].value).to.equal(81); // batt percentage
       expect(device.soilMoisture.service.characteristics[0].value).to.equal("Soil Moisture 49 %")
       done();
     });