+++ tests/unit/devices/WS68.spec.ts
@@ -25,7 +25,11 @@ configs.forEach(config => {
   describe(`WS68 device should be configured ${config}`, () => {
     before('Initialize device', () => {
       platform = createPlatform(config);
-      accessory = new api.platformAccessory('Accessory', "5746853e-4fee-4e47-97dd-53065ef1de03")
+      accessory = new api.platformAccessory('Accessory', "5746853e-4fee-4e47-97dd-53065ef1de03");
+
+      platform.config.nameOverrides = [];
+      platform.config.hidden = {};
+
       device = new WS68(platform, accessory);
     });
 
@@ -45,7 +49,7 @@ configs.forEach(config => {
       expect(device.windGust).to.not.be.undefined;
       expect(device.maxDailyGust).to.not.be.undefined;
 
-      expect(device.battery.displayName).to.equal('');
+      expect(device.battery.service.displayName).to.equal('Battery');
       expect(device.solarRadiation.service.characteristics[0].value).to.equal("Solar Radiation");
       expect(device.uvIndex.service.characteristics[0].value).to.equal("UV Index");
       expect(device.windDirection.service.characteristics[0].value).to.equal("Wind Direction");
@@ -60,7 +64,7 @@ configs.forEach(config => {
       device = new WS68(platform, accessory);
       device.update(dataReport);
 
-      expect(device.battery.characteristics[0].value).to.equal(0); // low batt
+      expect(device.battery.service.characteristics[1].value).to.equal(0); // low batt
       expect(device.solarRadiation.service.characteristics[0].value).to.equal("Solar Radiation 36890lx");
       expect(device.uvIndex.service.characteristics[0].value).to.equal("UV Index 1");
       expect(device.windDirection.service.characteristics[0].value).to.equal("Wind Direction 285° (W)");