+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.7.0
 
 * Add support for WN and WS series Ecowitt consoles (see README)
+* Add support for hiding battery service and setting name override for battery service
 * Consolidate gateway/console hiding feature to single hide/show setting
 * Internal refactor of data report request parsing and base station discovery